OLD ART website code/requirements/common.txt
Django==1.11.20
django-admin-honeypot==1.1.0
django_cron==0.5.1
django-ipware==2.1.0
django-model-utils==3.1.2
django-mptt==0.9.1
django-pipeline==1.6.14
django-redis==4.10.0
django-softhyphen==1.1.0
django-solo==1.1.3
django-suit==0.2.23
django-suit-ckeditor==0.0.2
ajax-views==0.2.2
beautifulsoup4==4.7.1
html5lib==1.0.1
csscompressor==0.9.5
jsmin==3.0.1
ipython==6.5.0
pillow==5.4.1
psycopg2-binary==2.8.4
pytils==0.3
pytz==2018.9
redis==3.5.3
requests==2.21.0
PyLinkedinAPI==0.1.5
facebook-sdk==3.1.0
python-twitter==3.5
django-recaptcha==2.0.4

MarkupSafe==2.0.1
jinja2==2.10.3
django_jinja==2.4.1

sendgrid==6.7.1  # or the latest compatible version

# pysyge==0.2.0
# premailer==3.1.1
# celery==3.1.23
# django-celery-email==1.1.4
